PEP 8 + pip install
1) Increase PEP 8 style
2) Change InstagramAPI\__init__.py
3) Move all from InstagramAPI\__init__.py to InstagramAPI\InstagramAPI.py
4) Now you can install using pip: `py -m install InstagramAPI` or `pip install InstagramAPI`
